fixed bug during calculation. int is not long enough for storage of device hours

+++ include/utils.h
@@ -33,7 +33,7 @@ class utils {
 		 */
 		static float FailureRate2FIT(float failureRate, float deviceHours);
 
-		static float calcFIT(float conflevel, int failures, int devhours);
+		static float calcFIT(float conflevel, int failures, double devhours);
 };
 
 #endif
